+++ text_data_detector/CHANGELOG.md
@@ -1,4 +1,8 @@
+## 0.1.0
+
+- Add `CalendarEventDetector` for dates and times.
+- Some improvements have been made to the phone detector.
+
 ## 0.0.1
 
 - Initial version.
-- Add opt-in `CalendarEventDetector` for dates and times.